- Create -18dbFS 1kHz test tone to make archive sample files better for testing - Declare function parameters as const when data is not modified - Fix archive file format by writing label signalling v210 coding for 10-bit uncompressed picture data to the GenericPictureEssenceDescriptor::PictureEssenceCoding property rather than to the FileDescriptor::Codec property - Added transfer characteristic and system item package metadata keys for D10 - Removed semi-colon from ends of MXF_TRACK_NUM defines to allow use as a function parameter Patches provided by Rob Lievaart  - Fix bug where parameters passed to MSVC gmtime_s were the wrong way round  - Remove unused variable in mxf_get_timestamp_now  - Added some explicit casts to avoid MSVC warnings  - Replaced while(1) with for(;;) to stop MSVC warnings  - Disabled MSVC warning about assignment in conditional in

+++ lib/mxf/mxf_partition.c
@@ -1,5 +1,5 @@
 /*
- * $Id: mxf_partition.c,v 1.5 2009/09/18 14:37:27 philipn Exp $
+ * $Id: mxf_partition.c,v 1.6 2010/02/12 13:46:27 philipn Exp $
  *
  * MXF file partitions
  *
@@ -708,7 +708,7 @@ int mxf_read_header_pp_kl_with_runin(MXFFile* mxfFile, mxfKey* key, uint8_t* lle
         runinCheckCount++;
         if (byte == ((const uint8_t*)(&g_PartitionPackPrefix_key))[keyCompareByte])
         {
-            keyPtr[keyCompareByte++] = byte;
+            keyPtr[keyCompareByte++] = (uint8_t)byte;
             if (keyCompareByte == 11)
             {
                 break;
@@ -735,7 +735,7 @@ int mxf_read_header_pp_kl_with_runin(MXFFile* mxfFile, mxfKey* key, uint8_t* lle
         {
             return 0;
         }
-        keyPtr[keyCompareByte] = byte;
+        keyPtr[keyCompareByte] = (uint8_t)byte;
     }
     
     if (!mxf_is_header_partition_pack(&k))
